requirements.txt
fastapi[all]==0.95.0
uvicorn==0.18.2
SQLAlchemy==2.0.38
psycopg2==2.9.5
pydantic==2.0.6
python-dotenv==1.0.0
